refactor: separated functions into modules
