+++ src/app/styles/main-page.css
@@ -152,9 +152,9 @@
 }
 
 .elem__item:hover {
-  -webkit-box-shadow: 0px 5px 10px 2px rgba(34, 60, 80, 0.9);
-  -moz-box-shadow: 0px 5px 10px 2px rgba(34, 60, 80, 0.9);
-  box-shadow: 0px 5px 10px 2px rgba(34, 60, 80, 0.9);
+  -webkit-box-shadow: 0px 5px 10px 2px rgba(34, 60, 80, 0.2);
+  -moz-box-shadow: 0px 5px 10px 2px rgba(34, 60, 80, 0.2);
+  box-shadow: 0px 5px 10px 2px rgba(34, 60, 80, 0.2);
 }
 
 .elem__item img {
@@ -178,26 +178,17 @@
 
 }
 
-.add__item {
+.add__item, .drop__item {
   position: relative;
   z-index: 5000;
     color: #3b3d40;
     font-weight: 600;
     transition: 0.5s;
 }
-.add__item:hover {
+.add__item:hover, .drop__item:hover {
     cursor: pointer;
     color: rgb(43, 156, 64);
     text-decoration: underline;
-}
-
-.drop__item {
-  color: rgb(102, 102, 102);
-  transition: 0.5s;
-}
-.drop__item:hover {
-  cursor: pointer;
-  color: rgb(43, 156, 64);
 }
 
 .multi-range__main {
